Fix bun-compile crash: keep json-schema.ts out of the module-commands chunk
The test_tgz bun-compiled binary crashed with 'ReferenceError:
__promiseAll is not defined'. Cause: module-commands.ts (a
dynamically-imported chunk) gained an import of getSchemaTypes from
json-schema.ts, whose top-level `await Promise.all([...])` validator
probing then got shared between the entry and dynamic chunks - and
bun's bundler emits its __promiseAll helper reference without defining
it in that shape (the base branch is green precisely because it lacks
this import edge).

getSchemaTypes is a pure function, so it moves to util.ts (which has no
top-level await) with json-schema.ts re-exporting it - no API change.

Verified bidirectionally with a local repro mirroring CI's
bun build --compile step: pre-fix dist crashes with CI's exact error,
fixed dist runs clean. tsc, lint and the full 369-test suite pass.
